FastAPI/requirements.txt
wheel==0.40.0
requests==2.28.2
fastapi==0.95.1
uvicorn[standard]==0.21.1
azure-storage-blob==12.16.0
azure-ai-textanalytics==5.2.1
prometheus-fastapi-instrumentator~=1.1.1
prometheus-client~=0.8.0
opentelemetry-distro==0.39b0
opentelemetry-exporter-otlp-proto-grpc==1.12.0
opentelemetry-instrumentation-fastapi==0.39b0